requirements.txt
asgiref==3.10.0
certifi==2025.10.5
charset-normalizer==3.4.3
dj-database-url==3.1.2
Django==5.2.7
django-environ==0.12.0
gunicorn==23.0.0
idna==3.10
joblib==1.5.2
numpy==2.3.3
nvidia-nccl-cu12==2.28.3
packaging==25.0
pandas==2.3.3
psycopg2-binary==2.9.11
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.5
scikit-learn==1.8.0
scipy==1.16.2
six==1.17.0
sqlparse==0.5.3
threadpoolctl==3.6.0
tzdata==2025.2
urllib3==2.5.0
whitenoise==6.12.0
xgboost==3.0.5
